+++ CHANGES.rst
@@ -9,6 +9,13 @@ Dependencies
 Incompatible changes
 --------------------
 
+* #13044: Remove the internal and undocumented ``has_equations`` data
+  from the :py:class:`!MathDomain`` domain.
+  The undocumented :py:meth:`!MathDomain.has_equations` method
+  now unconditionally returns ``True``.
+  These are replaced by the ``has_maths_elements`` key of the page context dict.
+  Patch by Adam Turner.
+
 Deprecated
 ----------